Fix sample names for samples in audiotable 2
Naming convention:
type-NAME-KEY-AUDIOTABLE

type: (e.g. 'inst' or 'perc' - keep to 4 characters max)
Used to denote 'inst' or 'perc' for a single set of conga samples currently. MM has two conga sets, one in use mainly as a percussion set in various banks, and another as an instrument in Woodfall's instrument bank.

NAME: (usually the instrument name the sample is from)
This is just the name of the sample.

KEY: (e.g. 'C060' or 'MUTE' - keep to 4 characters max)
For pitched samples this is the root key of the sample using note letter and MIDI numbering. For unpitched samples this is generally the type of hit or sound.

AUDIOTABLE: (e.g. 'AT02' - keep to 4 characters max)
The number of the audiotable the sample is located in. This is only used for samples not in audiotable 0. (Most likely not needed for decomp purposes as they're split already, but kept in for now)